+++ src/components/Map.vue
@@ -671,7 +671,7 @@ export default {
             if (this.$store.state.streamgageState == true) {
               let marker = L.marker([lat, lng], {
                 icon: this.nwisIcon,
-                zIndexOffset: 100, // add marker on top of other map layers
+                zIndexOffset: 1000, // add marker on top of other map layers
               }).addTo(this.streamgageMarkers);
               marker.data = { siteName: siteName, siteCode: siteID };
             }
@@ -1289,6 +1289,7 @@ export default {
                 this.activeSubtypes.push("path");
                 marker = L.marker([lat, lng], {
                   icon: aqIcon,
+                  zIndexOffset: 100
                 }).addTo(this.pathLayer);
                 this.pathLayer.addTo(this.aqMarkers);
               } else if (Name === "BANK") {
@@ -1298,6 +1299,7 @@ export default {
                 this.activeSubtypes.push("bank");
                 marker = L.marker([lat, lng], {
                   icon: aqIcon,
+                  zIndexOffset: 100
                 }).addTo(this.bankLayer);
                 this.bankLayer.addTo(this.aqMarkers);
               } else if (Name === "ROAD") {
@@ -1307,6 +1309,7 @@ export default {
                 this.activeSubtypes.push("road");
                 marker = L.marker([lat, lng], {
                   icon: aqIcon,
+                  zIndexOffset: 100
                 }).addTo(this.roadLayer);
                 this.roadLayer.addTo(this.aqMarkers);
               } else if (Name === "CHORD") {
@@ -1316,6 +1319,7 @@ export default {
                 this.activeSubtypes.push("bridgeRiskDiv");
                 marker = L.marker([lat, lng], {
                   icon: aqIcon,
+                  zIndexOffset: 100
                 }).addTo(this.bridgeRiskLayer);
                 this.bridgeRiskLayer.addTo(this.aqMarkers);
               } else if (Name === "FACILITY") {
@@ -1325,6 +1329,7 @@ export default {
                 this.activeSubtypes.push("facility");
                 marker = L.marker([lat, lng], {
                   icon: aqIcon,
+                  zIndexOffset: 100
                 }).addTo(this.facilityLayer);
                 this.facilityLayer.addTo(this.aqMarkers);
               } else if (Name === "DECK") {
@@ -1334,6 +1339,7 @@ export default {
                 this.activeSubtypes.push("bridgeFlooded");
                 marker = L.marker([lat, lng], {
                   icon: aqIcon,
+                  zIndexOffset: 100
                 }).addTo(this.bridgeFloodedLayer);
                 this.bridgeFloodedLayer.addTo(this.aqMarkers);
               } else if (Name === "BFE") {
@@ -1343,6 +1349,7 @@ export default {
                 this.activeSubtypes.push("bfe");
                 marker = L.marker([lat, lng], {
                   icon: aqIcon,
+                  zIndexOffset: 100
                 }).addTo(this.bfeLayer);
                 this.bfeLayer.addTo(this.aqMarkers);
               } else {
@@ -1352,6 +1359,7 @@ export default {
                 this.activeSubtypes.push("other");
                 marker = L.marker([lat, lng], {
                   icon: aqIcon,
+                  zIndexOffset: 100
                 }).addTo(this.otherLayer);
                 this.otherLayer.addTo(this.aqMarkers);
               }