+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Fix lat/long/zoom display
+- position of popups
 
 ## [v0.5.0](https://github.com/USGS-WiM/Thresholds/releases/tag/v0.5.0) - 9/1/2021