+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fix lat/long/zoom display
 - position of popups
+- Fixed layer order when toggling RP layer on/off
 
 ## [v0.5.0](https://github.com/USGS-WiM/Thresholds/releases/tag/v0.5.0) - 9/1/2021