+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+-
+
+### Changed
+
+-
+
+### Fixed
+
+-
+
+## [v0.6.0](https://github.com/USGS-WiM/Thresholds/releases/tag/v0.6.0) - 9/22/2021
+
+### Added
+
 - Add tooltips for active flooding
 - Added NOAA Tides and Current Stations supporting layer
 - Added ability to change time period